feat: move send and sendBatch into the same factory
