+++ src/flowcore/index.ts
@@ -2,5 +2,4 @@ export * from "./filehook"
 export * from "./redis-queue"
 export * from "./wait-for-predicate"
 export * from "./webhook"
-export * from "./webhook-batch"
 export * from "./metadata-webhook.factory"
